requirements.txt
Flask
pymongo
bcrypt
python-dotenv
Werkzeug
Flask-Login
bson
Flask-Mail
gunicorn